requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.10
django-cors-headers==3.3.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
fcm-django==0.3.4
googlemaps==4.4.1
idna==2.10
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyfcm==1.4.7
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
